Added support for fetching PFObjects and converting their JSON structure back into PFSerializable subclasses such as PFObject, PFFile, PFUser, etc.
This required altering the PFSerializable::fromJson() method to instead be a static method for each subclass that handles converting the JSON into a new heap object returned as a QVariant. This is a downside of having all our constructors and destructors protected forcing the user to use the static creation methods. The only alternative would be to have the PFSerializable class handle creating all the subclass instances, but that would seriously break encapsulation and would make the PFSerializable class more like a factory rather than a light-weight base class.