Refactored some important methods including the PFSerializable::className() to pfClassName and removed the PFSerializable::fromVariant() method.
These two were causing confusing naming collisions with the usage of our main objects such as PFObject, PFACL and PFUser. It was especially confusing when pulling objects out of a PFObject using the objectForKey method. The usage is now as follows:

QVariant variant = pfObjectInstance->objectForKey(key);
PFObjectPtr object = PFObject::objectFromVariant(variant);

Since the fromVariant method is no longer available, it's much more intuitive which method you are actually looking for.

Also modified the PFObject::parseClassName() method to be className() instead. It never made sense to create the object with className and then have to query for that same value using the parseClassName() method.

All the tests have been updated to accommodate the changes.