Extended the PFFile class to be able to check the cloud to make sure the url is actually valid.
This is very useful when considering the fact that running the "Clean Up Files" operation in the Parse settings for you app will wipe out all files that aren't contained within an object. If you are doing large file uploads and waiting to connect those PFFiles to PFObjects at a later time, then you need to be able to check if those files are still around when attaching them to the PFObject. This patch allows you to do so very quickly.